컴포넌트 이름 일부 수정, 구조분해할당 수정, 커스텀 훅  추가
